change in test double on int

+++ src/test/java/ru/itpark/SportMasterBonusServiceTest.java
@@ -10,22 +10,22 @@ class SportMasterBonusServiceTest {
     void calculateBonus() {
         {
             SportMasterBonusService sportMasterBonusService = new SportMasterBonusService();
-            double calculateBonus = sportMasterBonusService.calculateBonus(10_000, 6300, 200);
+            int calculateBonus = sportMasterBonusService.calculateBonus(10_000, 6300, 200);
             assertEquals(1_320, calculateBonus);
         }
         {
             SportMasterBonusService sportMasterBonusService = new SportMasterBonusService();
-            double calculateBonus = sportMasterBonusService.calculateBonus(25_000, 30_000, 600);
+            int calculateBonus = sportMasterBonusService.calculateBonus(25_000, 30_000, 600);
             assertEquals(4_450, calculateBonus);
         }
         {
             SportMasterBonusService sportMasterBonusService = new SportMasterBonusService();
-            double calculateBonus = sportMasterBonusService.calculateBonus(151_000, 80_000, 1_000);
+            int calculateBonus = sportMasterBonusService.calculateBonus(151_000, 80_000, 1_000);
             assertEquals(24_100, calculateBonus);
         }
         {
             SportMasterBonusService sportMasterBonusService = new SportMasterBonusService();
-            double   calculateBonus = sportMasterBonusService.calculateBonus(0, 1_000, 0);
+            int   calculateBonus = sportMasterBonusService.calculateBonus(0, 1_000, 0);
             assertEquals(50, calculateBonus);
         }
     }
